feat, fix, test, refactor by ardila

+++ demo/src/main/java/co/edu/udes/activity/backend/demo/services/AcademicRecordService.java
@@ -1,7 +1,7 @@
 package co.edu.udes.activity.backend.demo.services;
 
 import co.edu.udes.activity.backend.demo.models.AcademicRecord;
-import co.edu.udes.activity.backend.demo.repository.AcademicRecordRepository;
+import co.edu.udes.activity.backend.demo.repositories.AcademicRecordRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,16 +18,16 @@ public class AcademicRecordService {
         return academicRecordRepository.findAll();
     }
 
-    public Optional<AcademicRecord> getAcademicRecordById(int id) {
-        return academicRecordRepository.findById(id);
+    public Optional<AcademicRecord> getAcademicRecordById(long id) {
+        return academicRecordRepository.findById((long) id);
     }
 
     public AcademicRecord saveAcademicRecord(AcademicRecord academicRecord) {
         return academicRecordRepository.save(academicRecord);
     }
 
-    public AcademicRecord updateAcademicRecord(int id, AcademicRecord updatedAcademicRecord) {
-        return academicRecordRepository.findById(id).map(academicRecord -> {
+    public AcademicRecord updateAcademicRecord(long id, AcademicRecord updatedAcademicRecord) {
+        return academicRecordRepository.findById( id).map(academicRecord -> {
             academicRecord.setAcademicHistory(updatedAcademicRecord.getAcademicHistory());
             academicRecord.setGroup(updatedAcademicRecord.getGroup());
             academicRecord.setStudent(updatedAcademicRecord.getStudent());
@@ -35,9 +35,9 @@ public class AcademicRecordService {
         }).orElse(null);
     }
 
-    public boolean deleteAcademicRecord(int id) {
-        if (academicRecordRepository.existsById(id)) {
-            academicRecordRepository.deleteById(id);
+    public boolean deleteAcademicRecord(long id) {
+        if (academicRecordRepository.existsById( id)) {
+            academicRecordRepository.deleteById( id);
             return true;
         }
         return false;
