+++ demo/src/main/java/co/edu/udes/activity/backend/demo/services/GroupService.java
@@ -1,7 +1,7 @@
 package co.edu.udes.activity.backend.demo.services;
 
 import co.edu.udes.activity.backend.demo.models.Group;
-import co.edu.udes.activity.backend.demo.repository.GroupRepository;
+import co.edu.udes.activity.backend.demo.repositories.GroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,7 +18,7 @@ public class GroupService {
         return groupRepository.findAll();
     }
 
-    public Optional<Group> getGroupById(int id) {
+    public Optional<Group> getGroupById(long id) {
         return groupRepository.findById(id);
     }
 
@@ -26,7 +26,7 @@ public class GroupService {
         return groupRepository.save(group);
     }
 
-    public Group updateGroup(int id, Group updatedGroup) {
+    public Group updateGroup(long id, Group updatedGroup) {
         return groupRepository.findById(id).map(group -> {
             group.setStudentsAmount(updatedGroup.getStudentsAmount());
             group.setTeacher(updatedGroup.getTeacher());
@@ -35,7 +35,7 @@ public class GroupService {
         }).orElse(null);
     }
 
-    public boolean deleteGroup(int id) {
+    public boolean deleteGroup(long id) {
         if (groupRepository.existsById(id)) {
             groupRepository.deleteById(id);
             return true;
